% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement.function.R
\name{implement}
\alias{implement}
\title{A plugin function}
\usage{
implement(
  mRNA,
  miRNA.x,
  start.x,
  end.x,
  miRNA.y,
  start.y,
  end.y,
  mRNAexpression,
  miRNAexpression
)
}
\arguments{
\item{mRNA}{the mRNA}

\item{miRNA.x}{the miRNA x}

\item{start.x}{the start site of miRNA x binding on mRNA}

\item{end.x}{the end site of miRNA x binding on mRNA}

\item{miRNA.y}{the miRNA y}

\item{start.y}{the start site of miRNA y binding on mRNA}

\item{end.y}{the end site of miRNA binding on mRNA}

\item{mRNAexpression}{mRNA expression matrix}

\item{miRNAexpression}{miRNA expression matrix}
}
\value{
implement returns a result vector
}
\description{
A plugin function for evaluating binding interaction and information values
}
