% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.function.R
\name{sycomore}
\alias{sycomore}
\title{A comprehensive function for sycomore}
\usage{
sycomore(miRtarget, miRNAexpression, mRNAexpression)
}
\arguments{
\item{miRtarget}{miRNA target binding matrix}

\item{miRNAexpression}{miRNA expression matrix}

\item{mRNAexpression}{mRNA expression matrix}
}
\value{
sycomore returns a result in the form of a data frame
}
\description{
The sycomore function is to quantify the synergistic and competitive relationship between miRNAs.
}
\examples{
# load datasets
load(system.file("extdata/tcga.brca.testdata.Rdata", package="sycomore"))
miRNAexpression <- log2(miRNAexpression + 1)
mRNAexpression <- log2(mRNAexpression + 1)
# call function
sycomore(miRtarget, miRNAexpression, mRNAexpression)

}
